fix: overflow bug found on real Samsung S24 Ultra, reduce CommonButton height
Real production bug found via device smoke test on a real Android
device (Samsung S24 Ultra, density override 480 + font_scale 1.08 —
narrower effective width than any device tested before): the
CurrencyCounter demo row (CurrencyCounter + two fixed-width
CommonButtons in a plain Row(mainAxisSize.min)) overflowed by several
pixels — reproducible live via `flutter run`, and the same overflow
cascaded into 2 unrelated tests failing later in the same continuous
integration_test binary (SoundToggleFab, ShimmerPlaceholder), not
separate bugs. Fixed by switching that Row to a Wrap, which flows
content to a new line instead of overflowing, no matter how narrow the
available width gets. Re-ran on the real device afterward: 4 passed +
4 documented skips, 0 failures — matching baseline exactly.

Also reduced CommonButton's vertical padding (16 -> 8 logical pixels
per side, s16 -> s8) per explicit feedback that the pill buttons read
as too tall — height drops from 63px to a more standard ~44-48px range.
Updated the two ShopItemCard golden images that captured its old
proportions.

Added regression tests for both: an isolated Wrap-at-296px-width test
(the exact narrow constraint recorded from the live device error) for
the overflow fix, and a height-under-52px assertion for the button
sizing.

Verified: flutter analyze clean (root + example), full suite green
(382 root + 27 example tests, 0 regressions), device smoke test on the
real Samsung S24 Ultra (not a simulator) confirming both fixes visually
and via the integration test suite.

+++ lib/presentation/widgets/common/common_button.dart
@@ -70,7 +70,7 @@ class CommonButton extends StatelessWidget {
     final outlined = variant == CommonButtonVariant.secondary;
     return Container(
       width: width ?? 240,
-      padding: const EdgeInsets.symmetric(vertical: NeonTheme.s16),
+      padding: const EdgeInsets.symmetric(vertical: NeonTheme.s8),
       decoration: BoxDecoration(
         gradient: outlined
             ? null
